requirements.txt
# Core dependencies
crewai>=0.134.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.20

# LLM and AI
langchain-openai>=0.3.0
openai>=1.93.0
langchain-community>=0.0.20
langchain-core>=0.3.0

# PDF processing
pypdf>=5.6.0
pdfplumber>=0.11.0

# Environment and utilities
python-dotenv>=1.0.0
pydantic>=2.4.0

# Optional: For enhanced features
# redis>=4.0.0  # For queue worker model
# celery>=5.0.0  # For background tasks
# sqlalchemy>=2.0.0  # For database integration 